move filecoin options to @ganache/filecoin-options
